Redo csc calculations
There is no possibility to forward the csc matrix to linux-sunxi kernel
display driver. We have to set the values for brightness, saturation,
contrast and hue via ioctl. But there are no attributes for these values
in VDPAU to get them directly. So we need some reverse calculation to
get these values out of the (changed) csc matrix.
The old calculation was wrong and is now corrected to get the right
values for brightness, saturation, contrast and hue.

Remember that there is a kernel bug in linux-sunxi kernel-3.4 at time
of commiting. Use the attached patch to fix this problem.